# docs/设计/接口契约/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campus-Link API",
    "description": "重庆工程学院计算机专业学生交流论坛（MVP）。错误码分段见技术方案第 5 节。",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "我的主页（F-ACC-002 最小版；他人主页在后续 Sprint）",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserVo"
                }
              }
            }
          },
          "401": {
            "description": "未登录（4001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在（2007）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "系统繁忙，请稍后再试（9999）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "修改我的资料（需登录）：昵称 / 专业 / 签名；字段传 null 即不改、传空串即清空（昵称除外）；全量回显 UserVo",
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserVo"
                }
              }
            }
          },
          "400": {
            "description": "参数错误（1001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "未登录（4001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在（2007）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "账号已被封禁（2006）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "资料修改过于频繁，请稍后再试（2008）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "系统繁忙，请稍后再试（9999）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/posts/following": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "「关注」Feed（需登录）：本人关注作者的时间序帖子；未关注任何人时为空页（CR-074）",
        "operationId": "following",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageVoPostSummaryVo"
                }
              }
            }
          },
          "400": {
            "description": "参数错误（1001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "未登录（4001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "系统繁忙，请稍后再试（9999）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateProfileCommand": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2
          },
          "major": {
            "type": "string",
            "maxLength": 64,
            "minLength": 0,
            "pattern": "^[^<>\\p{Cntrl}\\u200B\\u200C\\u200D\\u2060\\uFEFF]*$"
          },
          "bio": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0,
            "pattern": "^[^<>\\p{Cntrl}\\u200B\\u200C\\u200D\\u2060\\uFEFF]*$"
          }
        }
      },
      "ApiResponseUserVo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserVo"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "UserVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nickname": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "school": {
            "type": "string"
          },
          "major": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponsePageVoPostSummaryVo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageVoPostSummaryVo"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "PageVoPostSummaryVo": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostSummaryVo"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PostSummaryVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "boardCode": {
            "type": "string"
          },
          "boardName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "authorNickname": {
            "type": "string"
          },
          "replyCount": {
            "type": "integer",
            "format": "int32"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "viewCount": {
            "type": "integer",
            "format": "int32"
          },
          "coverUrl": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "accepted": {
            "type": "boolean"
          }
        }
      },
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "登录返回的 JWT，请求头 `Authorization: Bearer <token>`。框架在进入业务代码前按端点注解裁决：标 `@SecurityRequirement` 的端点要求已登录，标 `@PublicEndpoint` 的放行，module 端点两者都没标即 fail-closed；未登录由过滤器链直接产出 401 / 4001（CR-031，SecurityConfig 已无 permitAll 兜底）。角色与资源级授权仍归业务代码的 CurrentUser（requireRole → 403 / 4002）；受保护端点必须声明本 security 并真的调用 CurrentUser，该纪律由 ArchitectureGuardTest 机器校验。",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}